Update HL1MDLLoader.cpp
Fix possible x64 issue.

+++ code/MDL/HalfLife/HL1MDLLoader.cpp
@@ -367,7 +367,7 @@ void HL1MDLLoader::read_texture(const Texture_HL1 *ptexture,
     for (outwidth = 1; outwidth < ptexture->width; outwidth <<= 1)
         ;
 
-    if ((size_t)outwidth > BuffenLen) {
+    if ( outwidth > static_cast<int>(BuffenLen)) {
         outwidth = BuffenLen;
     }
